Migrate allegany (rotated-SOVC PDF, two side-by-side contests)
